[#9] musicPlayer 컴포넌트 분리
