+++ frontend/src/components/Room/MusicPlayer/Title.tsx
@@ -1,0 +1,10 @@
+function Title({ name, singer }: { name: string | undefined; singer: string | undefined }) {
+  return (
+    <div className="musicplayer-title-area">
+      <span className="musicplayer-title">{name}</span>
+      <span className="musicplayer-subtitle">{singer}</span>
+    </div>
+  );
+}
+
+export default Title;
